requirements.txt
python-telegram-bot==21.8
sqlalchemy==2.0.36
python-dotenv~=1.0.1